레벨, 시간에 따른 스탯 변화 추가
플레이어는 레벨 비례 공, 체 증가
몬스터는 시간 비례 공, 체 증가

+++ Assets/0.Scripts/MonsterScripts/MonsterCtrl.cs
@@ -37,7 +37,7 @@ public class MonsterCtrl : MonoBehaviour
     }
 
     //몬스터가 피해를 받는 함수
-    public void Damaged(int val)
+    public void Damaged(float val)
     {
         mon.monCurHP -= val;
     }
